requirements.txt
#### **`requirements.txt`**
plaintext
python-telegram-bot==20.5
openai==0.27.8